airflow/requirements.txt
boto3
python-dotenv
pandas
pyarrow
fastparquet
requests
